fix: consolidate GET /trust and GET /trust/status into one endpoint (#609)
GET /trust (IBasicTrust) and GET /trust/status (ITrustStatus) were two
auth-only, unfiltered list views of the same resource, with ITrustStatus a
strict superset of IBasicTrust. Collapse them onto a single GET /trust that
returns the richer shape (region, last_heartbeat, project_count).

flip-api:
- get_trusts now returns list[ITrustStatus], folding in the _as_utc_iso helper
  and the per-trust project-count GROUP BY from the deleted get_trust_statuses,
  plus its generic-500 error handling (no exception-text leak).
- Drop IBasicTrust and the get_trust_statuses router registration.

flip-ui:
- trust-service exposes one ITrustResponse (6-field superset) and two thin
  wrappers over GET /trust: getTrusts (swallow -> [] for the pickers) and
  getTrustStatuses (propagate so the Connection Status loader holds).
- ConnectionStatus.vue points at the consolidated endpoint under its own SWRV
  cache key, keeping its 15s poll independent of the app-wide /trust bootstrap.
- Remove trust-status-service.

Tests merged/updated accordingly; the #557 non-admin-access regression now
asserts against GET /trust.
