+++ flip-api/src/flip_api/domain/interfaces/trust.py
@@ -20,19 +20,14 @@ from flip_api.domain.schemas.users import CognitoUser
 # Interfaces
 
 
-class IBasicTrust(BaseModel):
-    id: UUID
-    name: str
-    code: str | None = None
-
-
 class ITrustStatus(BaseModel):
-    """Connection-status view of a trust, readable by any authenticated user.
+    """Trust list item with connection status, readable by any authenticated user.
 
-    Powers the Connection Status page (trust table + topology). Every field is
-    benign trust metadata — no secrets — so the endpoint that returns this
-    (GET /trust/status) is intentionally not admin-gated. Creating a trust
-    (POST /admin/trusts) stays admin-only.
+    Backs ``GET /trust`` — the single list-of-trusts endpoint — powering both the
+    trust pickers (project staging, cohort query, charts) and the Connection
+    Status page (trust table + topology). Every field is benign trust metadata —
+    no secrets — so the endpoint is intentionally not admin-gated. Creating a
+    trust (``POST /admin/trusts``) stays admin-only.
     """
 
     id: UUID
